requirements.txt
flask
psycopg2-binary
requests
python-dotenv
langchain
langchain-groq
gunicorn